test: move tests to apps-test file
